refactor(Timeline): Move to partials
Consistent pattern across components
