+++ packages/react-component-library/src/components/Timeline/TimelineMonth.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { format, endOfMonth, differenceInDays, min, max } from 'date-fns'
-import styled from 'styled-components'
-import { selectors } from '@royalnavy/design-tokens'
 
 import { DATE_MONTH_FORMAT } from './constants'
 import { formatPx } from './helpers'
+import { StyledMonth } from './partials/StyledMonth'
+import { StyledMonthTitle } from './partials/StyledMonthTitle'
 import { TimelineDay } from './context/types'
 
 interface TimelineMonthProps {
@@ -20,38 +20,6 @@ interface TimelineMonthProps {
   startDate: Date
 }
 
-const { spacing, color, zIndex, fontSize } = selectors
-
-const StyledTimelineMonth = styled.div`
-  position: relative;
-  display: inline-flex;
-  align-items: center;
-  height: 4rem;
-
-  &:last-of-type {
-    border-right: none;
-  }
-
-  &::after {
-    position: absolute;
-    right: 0;
-    top: 0;
-    content: '';
-    display: inline-block;
-    width: 1rem;
-    height: 100vh;
-    border-right: ${spacing('px')} dashed ${color('neutral', '200')};
-    z-index: ${zIndex('body', 1)};
-  }
-`
-
-const StyledTitle = styled.span`
-  font-size: ${fontSize('xl')};
-  font-weight: 600;
-  color: ${color('neutral', '600')};
-  padding-left: ${spacing('8')};
-`
-
 function renderDefault({
   dayWidth,
   daysTotal,
@@ -62,14 +30,16 @@ function renderDefault({
   startDate: Date
 }): React.ReactElement {
   return (
-    <StyledTimelineMonth
+    <StyledMonth
       style={{
         width: formatPx(dayWidth, daysTotal),
       }}
       data-testid="timeline-month"
     >
-      <StyledTitle>{format(startDate, DATE_MONTH_FORMAT)}</StyledTitle>
-    </StyledTimelineMonth>
+      <StyledMonthTitle>
+        {format(startDate, DATE_MONTH_FORMAT)}
+      </StyledMonthTitle>
+    </StyledMonth>
   )
 }
