+++ packages/react-component-library/src/components/Timeline/TimelineTodayMarker.tsx
@@ -1,10 +1,10 @@
 import React, { useContext } from 'react'
-import styled from 'styled-components'
-import { selectors } from '@royalnavy/design-tokens'
 
 import { ComponentWithClass } from '../../common/ComponentWithClass'
+import { StyledTodayMarker } from './partials/StyledTodayMarker'
+import { StyledTodayMarkerWrapper } from './partials/StyledTodayMarkerWrapper'
 import { TimelineContext } from './context'
-import { useTimelinePosition } from './hooks/useTimelinePosition'
+import { useTimelinePosition } from './hooks'
 
 export interface TimelineTodayMarkerWithRenderContentProps
   extends ComponentWithClass {
@@ -20,33 +20,9 @@ export type TimelineTodayMarkerProps =
   | TimelineTodayMarkerWithRenderContentProps
   | TimelineTodayMarkerWithChildrenProps
 
-const { spacing, color, zIndex, fontSize } = selectors
-
-const StyledTimelineTodayMarkerWrapper = styled.div`
-  position: relative;
-`
-
-const StyledTimelineTodayMarker = styled.div`
-  position: absolute;
-  top: 4rem;
-  display: inline-block;
-  width: ${spacing('px')};
-  height: 100vh;
-  background-color: ${color('danger', '500')};
-  z-index: ${zIndex('body', 999)};
-
-  &::before {
-    content: 'Today';
-    display: inline-block;
-    color: ${color('danger', '500')};
-    font-size: ${fontSize('s')};
-    transform: translate(-50%, -175%);
-  }
-`
-
 function renderDefault({ offset }: { offset: string }) {
   return (
-    <StyledTimelineTodayMarker
+    <StyledTodayMarker
       data-testid="timeline-today-marker"
       style={{ left: offset }}
     />
@@ -69,13 +45,13 @@ export const TimelineTodayMarker: React.FC<TimelineTodayMarkerProps> = ({
   if (startsBeforeStart || startsAfterEnd) return null
 
   return (
-    <StyledTimelineTodayMarkerWrapper
+    <StyledTodayMarkerWrapper
       data-testid="timeline-today-marker-wrapper"
       role="presentation"
       {...rest}
     >
       {render ? render(today, offset) : renderDefault({ offset })}
-    </StyledTimelineTodayMarkerWrapper>
+    </StyledTodayMarkerWrapper>
   )
 }
