+++ packages/react-component-library/src/components/Timeline/TimelineWeek.tsx
@@ -1,17 +1,14 @@
 import React from 'react'
-import styled from 'styled-components'
 import { format, differenceInDays, endOfWeek, max, min } from 'date-fns'
-import { selectors } from '@royalnavy/design-tokens'
 
 import {
   ACCESSIBLE_DATE_FORMAT,
   DATE_WEEK_FORMAT,
   WEEK_START,
-  TIMELINE_BG_COLOR,
-  TIMELINE_ALT_BG_COLOR,
-  TIMELINE_BORDER_COLOR,
 } from './constants'
 import { formatPx, isOdd } from './helpers'
+import { StyledWeekTitle } from './partials/StyledWeekTitle'
+import { StyledWeek } from './partials/StyledWeek'
 import { TimelineDay } from './context/types'
 
 interface TimelineWeekProps {
@@ -30,37 +27,6 @@ interface TimelineWeekProps {
   startDate: Date
 }
 
-const { spacing, fontSize, color, zIndex } = selectors
-
-interface StyledTimelineWeekProps {
-  isOddNumber: boolean
-  marginLeft: string
-  width: string
-}
-
-const StyledTimelineWeek = styled.div<StyledTimelineWeekProps>`
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  height: 2.5rem;
-  background-color: ${({ isOddNumber }) =>
-    isOddNumber ? TIMELINE_ALT_BG_COLOR : TIMELINE_BG_COLOR};
-  border-top: ${spacing('px')} solid ${TIMELINE_BORDER_COLOR};
-  border-bottom: ${spacing('px')} solid ${TIMELINE_BORDER_COLOR};
-  justify-content: unset;
-  margin-left: ${({ marginLeft }) => marginLeft};
-  width: ${({ width }) => width};
-`
-
-const StyledTitle = styled.span`
-  font-weight: 600;
-  font-size: ${fontSize('m')};
-  color: ${color('neutral', '400')};
-  background-color: inherit;
-  z-index: ${zIndex('body', 2)};
-  margin-left: ${spacing('4')};
-`
-
 function renderDefault({
   isOddNumber,
   offsetPx,
@@ -73,14 +39,14 @@ function renderDefault({
   startDate: Date
 }) {
   return (
-    <StyledTimelineWeek
+    <StyledWeek
       isOddNumber={isOddNumber}
       marginLeft={offsetPx}
       width={widthPx}
       data-testid="timeline-week"
     >
-      <StyledTitle>{format(startDate, DATE_WEEK_FORMAT)}</StyledTitle>
-    </StyledTimelineWeek>
+      <StyledWeekTitle>{format(startDate, DATE_WEEK_FORMAT)}</StyledWeekTitle>
+    </StyledWeek>
   )
 }
