+++ packages/react-component-library/src/components/Timeline/partials/StyledDays.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const StyledDays = styled.div`
+  white-space: nowrap;
+`
