+++ packages/react-component-library/src/components/Timeline/partials/StyledHeader.tsx
@@ -1,0 +1,3 @@
+import styled from 'styled-components'
+
+export const StyledHeader = styled.div``
