+++ packages/react-component-library/src/components/Timeline/partials/StyledHeaderRow.tsx
@@ -1,0 +1,11 @@
+import styled, { css } from 'styled-components'
+
+import { TimelineRow } from '../TimelineRow'
+
+export const StyledHeaderRow = styled<any>(TimelineRow)`
+  ${({ isShort }) =>
+    isShort &&
+    css`
+      height: 2.5rem;
+    `}
+`
