+++ packages/react-component-library/src/components/Timeline/partials/StyledHours.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const StyledHours = styled.div`
+  white-space: nowrap;
+`
