+++ packages/react-component-library/src/components/Timeline/partials/StyledMonths.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const StyledMonths = styled.div`
+  white-space: nowrap;
+`
