+++ packages/react-component-library/src/components/Timeline/partials/StyledNoEvents.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const StyledNoEvents = styled.span`
+  display: none;
+`
