+++ packages/react-component-library/src/components/Timeline/partials/StyledRow.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+
+export const StyledRow = styled.div`
+  display: flex;
+  height: 4rem;
+`
