+++ packages/react-component-library/src/components/Timeline/partials/StyledRowContent.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const StyledRowContent = styled.div`
+  position: relative;
+`
