+++ packages/react-component-library/src/components/Timeline/partials/StyledRowWeeks.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+export const StyledRowWeeks = styled.div`
+  position: absolute;
+  top: 0;
+  left: 0;
+  white-space: nowrap;
+`
