+++ packages/react-component-library/src/components/Timeline/partials/StyledRowWeeksWrapper.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const StyledRowWeeksWrapper = styled.div`
+  position: relative;
+`
