+++ packages/react-component-library/src/components/Timeline/partials/StyledRows.tsx
@@ -1,0 +1,15 @@
+import styled, { css } from 'styled-components'
+
+interface StyledRowsProps {
+  defaultStyles: boolean
+}
+
+export const StyledRows = styled.div<StyledRowsProps>`
+  ${({ defaultStyles }) =>
+    defaultStyles &&
+    css`
+      width: auto;
+      height: auto;
+      min-height: 4rem;
+    `}
+`
