+++ packages/react-component-library/src/components/Timeline/partials/StyledTimeline.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+import { selectors } from '@royalnavy/design-tokens'
+
+const { color, zIndex } = selectors
+
+export const StyledTimeline = styled.div`
+  position: relative;
+  z-index: ${zIndex('timeline', 0)};
+  display: flex;
+  width: 100%;
+  background-color: ${color('neutral', 'white')};
+`
