+++ packages/react-component-library/src/components/Timeline/partials/StyledTodayMarkerWrapper.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const StyledTodayMarkerWrapper = styled.div`
+  position: relative;
+`
