+++ packages/react-component-library/src/components/Timeline/partials/StyledWeeks.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const StyledWeeks = styled.div`
+  white-space: nowrap;
+`
